tools/batch: Add an option to limit the number of goroutines to spawn.
Spawning too many goroutines may rapidly tear down a server due to the
fact that too many file descriptor may be open at once.

Moreover, this may also create too many connections to the database.

Hence, add an option to limit the number of goroutines to spawn.
The bad part about it is that when the max number of goroutines has been
spawned, it waits for every goroutine to finish before continuing.
Due to this, server resources are not fully used all the time.
However, improving the situation may also mean using locking primitives
and so on and complicate the code.
We'll see later if that could be worth it or not.