misc: Add an option to retrieve commit patches.
Add a field to the configuration file that allows specifying whether to
retrieve commit patches or not. If set to true, commit patches will be
added to a commit.
Due to the fact that the data produced may be huge, do not add this kind
of data into the database. Hence, this option is only valid when
outputting JSON objects.