misc: Add support for commit deltas.
Add a new option to the configuration file to include diff deltas of a
commit.

In order to do this, the patches field of a commit has been removed and
placed inside a diff delta field. Hence, a commit now includes a slice
of diff deltas, each delta being a delta against one file touched by the
commit.

The option that allows keeping patches or not remains. If set to false,
the patch field of a diff delta structure will simply not be set.

While there, also document all models with more details.