db: expand unique constraint from hash to hash + vcs_id (close #1)
In the process, make the vcs_id column not null. A VCS that does not
have a way to uniquely identify a commit within a repository is not
likely to exist.

The probability that both a hash and a vcs_id are not unique together is
/very/ unlikely. Hence, we avoid dropping commits now and if we ever see
a commit not being inserted because of a unique constraint violation,
then it will be time to rethink the way we prevent the same commit from
being inserted more than once into the database.