build: do not follow git2go 'next'
git2go used to provide scripts to statically compile libgit2 with
git2go.

However, this process is only available in the 'next' branch, which is
unstable. The 'master' branch, which follows latest release of libgit2,
does not provide any script to statically build libgit2 with git2go.

Hence, stop building libgit2 statically and following 'next' for now
until a proper solution is found.