build/travis: do not build with Go < 1.4
If we do not use 'go get -f ...' to install devmine related packages, it
fails to install due to the clone being from the git protocol rather
than the https protocol. Hence, in this case, building fail for Go >=
1.4.

On the other hand, if we use 'go get -f ...', then it fails to build on
Go versions older than 1.4 since the '-f' option did not exist.

Since we do not really care about older versions, stop build for Go
versions older than 1.4.

This does not mean that repotool cannot be built with older Go versions.
In fact, 'go get github.com/DevMine/repotool' will probably work for all
Go versions. We just do not actively check that it does.