repo: make use of tmp_dir_file_size_limit option
While here, fix issues about temporary directory removal.
To analyze a git repository, when in a tar archive format, repotool has
to extract the archive in order to process. Only in this case we need to
tell gitrepo that the git directory is a temporary directory, hence it
can take care of removing it after processing.