db: store users in hashmap to avoid db queries
It is not a surprise that this brings a huge speedup since user IDs were
queried for /each commit/ of the repository before this change. Now,
all user IDs are queried once and stored in a hashmap so access is O(1)
and we avoid querying the database.

Locally, on my laptop with a small database and a small repository with
only 473 commits, inserting the commits and commit deltas into the
database now takes ~250ms where it used to take ~5.7s.